Ignore response exceptions when from http client
Yada supports setting keys in the ex-data of an exception.  In
particular, :status and :body.  A problem is that it is also produced by
Clojure http clients.  Usually this is not what the user intends, so it
can accidentally leak unhandled failures from external APIs which may
contain sensitive information.  The goal is to ignore these exceptions,
as is best possible.  Users can restore previous behavior by rewrapping
the ex-data with their own exception with only the required keys.

Detection is based on keys which are found in aleph, hato, clj-http, and
clj-http-lite.  These are the most popular libraries for Clojure.